+++ game-mods/mycu_external_app/ui/config.lua
@@ -1,3 +1,3 @@
---- Configuration file for X4 External App
+--- Configuration file for X4 Dashboard Bridge
 host = '127.0.0.1'
 port = 3001